updated main class, irrtopic class, me class
IMPORTANT:
! PLEASE ADD A return; TO ALL runTopic() METHODS in ALL topic classes -
this is so the program knows when to go back to the method it was called
from.
! PLEASE USE runTopic(String input) AS YOUR MAIN CONVERSATION METHOD! -
this is so irrelavantTopic class is able to run the respective class.
You can refer to the me class to see what I mean and test the program.

Additionally, the program runs all topics at once. we will decide on
individual implementation later.